CNN-FOX News/requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
newsapi-python==0.2.6
requests==2.23.0
six==1.12.0
urllib3==1.25.8
